requirements.txt
attrs==21.2.0
beautifulsoup4==4.10.0
black==21.9b0
click==8.0.1
coverage==5.5
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mypy-extensions==0.4.3
numpy==1.21.2
packaging==21.0
pandas==1.3.3
pathspec==0.9.0
platformdirs==2.3.0
pluggy==1.0.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
regex==2021.8.28
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.23
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.2
Werkzeug==2.0.1
